src/requirements.txt
json5==0.9.5
labelme==5.0.0
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.1.2.30
pandas==1.1.4
pathlib==1.0.1
Pillow==8.0.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
seaborn==0.11.0
tensorboard==2.4.1
tqdm==4.48.0
xlwt==1.3.0